+++ ecosystem/typescript/sdk/src/account/aptos_account.ts
@@ -5,12 +5,12 @@ import nacl from "tweetnacl";
 import * as bip39 from "@scure/bip39";
 import { bytesToHex } from "@noble/hashes/utils";
 import { sha3_256 as sha3Hash } from "@noble/hashes/sha3";
-import { derivePath } from "./utils/hd-key";
-import { HexString, MaybeHexString } from "./hex_string";
-import * as Gen from "./generated/index";
-import { Memoize } from "./utils";
-import { AccountAddress, AuthenticationKey, Ed25519PublicKey } from "./aptos_types";
-import { bcsToBytes } from "./bcs";
+import { derivePath } from "../utils/hd-key";
+import { HexString, MaybeHexString } from "../hex_string";
+import * as Gen from "../generated/index";
+import { Memoize } from "../utils";
+import { AccountAddress, AuthenticationKey, Ed25519PublicKey } from "../aptos_types";
+import { bcsToBytes } from "../bcs";
 
 export interface AptosAccountObject {
   address?: Gen.HexEncodedBytes;
