+++ ecosystem/typescript/sdk/src/account/index.ts
@@ -1,0 +1,2 @@
+export * from "./aptos_account";
+export * from "./account_address";
