+++ ecosystem/typescript/sdk/src/aptos_types/authenticator.ts
@@ -3,7 +3,7 @@
 
 /* eslint-disable @typescript-eslint/naming-convention */
 import { Serializer, Deserializer, Seq, deserializeVector, serializeVector } from "../bcs";
-import { AccountAddress } from "./account_address";
+import { AccountAddress } from "../account";
 import { Ed25519PublicKey, Ed25519Signature } from "./ed25519";
 import { MultiEd25519PublicKey, MultiEd25519Signature } from "./multi_ed25519";
