+++ ecosystem/typescript/sdk/src/aptos_types/rotation_proof_challenge.ts
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { Serializer } from "../bcs";
-import { AccountAddress } from "./account_address";
+import { AccountAddress } from "../account";
 
 export class RotationProofChallenge {
   constructor(
