+++ ecosystem/typescript/sdk/src/aptos_types/transaction.ts
@@ -21,7 +21,7 @@ import {
   Uint16,
   Uint256,
 } from "../bcs";
-import { AccountAddress } from "./account_address";
+import { AccountAddress } from "../account";
 import { TransactionAuthenticator } from "./authenticator";
 import { Identifier } from "./identifier";
 import { TypeTag } from "./type_tag";
