+++ ecosystem/typescript/sdk/src/tests/e2e/aptos_client.test.ts
@@ -3,15 +3,15 @@
 
 import { AptosClient } from "../../providers/aptos_client";
 import * as Gen from "../../generated/index";
-import { AptosAccount } from "../../aptos_account";
+import { AptosAccount } from "../../account/aptos_account";
 import {
   TxnBuilderTypes,
   TransactionBuilderMultiEd25519,
   TransactionBuilderRemoteABI,
 } from "../../transaction_builder";
-import { TokenClient } from "../../token_client";
+import { TokenClient } from "../../plugins/token_client";
 import { HexString } from "../../hex_string";
-import { getFaucetClient, longTestTimeout, NODE_URL } from "../../utils/test_helper.test";
+import { getFaucetClient, longTestTimeout, NODE_URL } from "../unit/test_helper.test";
 import { bcsSerializeUint64, bcsToBytes } from "../../bcs";
 import { Ed25519PublicKey } from "../../aptos_types";
