+++ ecosystem/typescript/sdk/src/tests/e2e/indexer.test.ts
@@ -1,10 +1,10 @@
-import { AptosAccount } from "../../aptos_account";
+import { AptosAccount } from "../../account/aptos_account";
 import { AptosClient } from "../../providers/aptos_client";
 import { bcsSerializeBool } from "../../bcs";
 import { FaucetClient } from "../../providers/faucet_client";
 import { IndexerClient } from "../../providers/indexer";
-import { TokenClient } from "../../token_client";
-import { API_TOKEN, longTestTimeout } from "../../utils/test_helper.test";
+import { TokenClient } from "../../plugins/token_client";
+import { API_TOKEN, longTestTimeout } from "../unit/test_helper.test";
 import { Network, NetworkToIndexerAPI, NetworkToNodeAPI, sleep } from "../../utils";
 
 describe("Indexer", () => {
