+++ ecosystem/typescript/sdk/src/tests/unit/deserializer.test.ts
@@ -1,7 +1,7 @@
 // Copyright © Aptos Foundation
 // SPDX-License-Identifier: Apache-2.0
 
-import { Deserializer } from "./deserializer";
+import { Deserializer } from "../../bcs/deserializer";
 
 describe("BCS Deserializer", () => {
   it("deserializes a non-empty string", () => {
