+++ ecosystem/typescript/sdk/src/tests/unit/helper.test.ts
@@ -1,8 +1,8 @@
 // Copyright © Aptos Foundation
 // SPDX-License-Identifier: Apache-2.0
 
-import { AccountAddress } from "../aptos_types";
-import { Deserializer } from "./deserializer";
+import { AccountAddress } from "../../aptos_types";
+import { Deserializer } from "../../bcs/deserializer";
 import {
   bcsSerializeBool,
   bcsSerializeBytes,
@@ -17,8 +17,8 @@ import {
   deserializeVector,
   serializeVector,
   serializeVectorWithFunc,
-} from "./helper";
-import { Serializer } from "./serializer";
+} from "../../bcs/helper";
+import { Serializer } from "../../bcs/serializer";
 
 test("serializes and deserializes a vector of serializables", () => {
   const address0 = AccountAddress.fromHex("0x1");
