+++ ecosystem/typescript/sdk/src/tests/unit/misc.test.ts
@@ -1,7 +1,7 @@
 // Copyright © Aptos Foundation
 // SPDX-License-Identifier: Apache-2.0
 
-import { AptosClient } from "../providers/aptos_client";
+import { AptosClient } from "../../providers/aptos_client";
 
 test("test fixNodeUrl", () => {
   expect(new AptosClient("https://test.com").client.request.config.BASE).toBe("https://test.com/v1");
