+++ ecosystem/typescript/sdk/src/tests/unit/property_map_serde.test.ts
@@ -1,4 +1,4 @@
-import { deserializeValueBasedOnTypeTag, getPropertyType, getPropertyValueRaw } from "./property_map_serde";
+import { deserializeValueBasedOnTypeTag, getPropertyType, getPropertyValueRaw } from "../../utils/property_map_serde";
 import {
   bcsSerializeBool,
   bcsSerializeStr,
@@ -7,10 +7,10 @@ import {
   bcsSerializeUint64,
   bcsToBytes,
   Bytes,
-} from "../bcs";
-import { AccountAddress } from "../aptos_types";
-import { HexString } from "../hex_string";
-import { TypeTagParser } from "../transaction_builder";
+} from "../../bcs";
+import { AccountAddress } from "../../aptos_types";
+import { HexString } from "../../hex_string";
+import { TypeTagParser } from "../../transaction_builder";
 
 test("test property_map_serializer", () => {
   function isSame(array1: Bytes, array2: Bytes): boolean {
